% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-utils.R
\name{list_functions}
\alias{list_functions}
\title{List functions defined in a package}
\usage{
list_functions(package)
}
\arguments{
\item{package}{Character. Package name}
}
\value{
Character vector of function names defined in this package
}
\description{
Returns only functions actually defined in the package, excluding imports and re-exports
}
\examples{
list_functions("MomX")

}
