% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX-package.R
\docType{package}
\name{MomX-package}
\alias{MomX}
\alias{MomX-package}
\title{MomX: Easily Install and Load MomX packages}
\description{
MomX packages are an ecosystem of packages for
everything 2D morphometrics, that is the statistical description
of shape and its (co)variation.
MomX packages share common principles and work together well.
This eponymous package is designed to make it easy to install
and load MomX packages in a single step.
Learn more about MomX at <https://momx.github.io/MomX/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://momx.github.io/MomX}
  \item Report bugs at \url{https://github.com/MomX/MomX}
}

}
\author{
\strong{Maintainer}: Vincent Bonhomme \email{bonhomme.vincent@gmail.com}

}
\keyword{internal}
