% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{MomX_conflicts}
\alias{MomX_conflicts}
\title{Conflicts between MomX and other packages}
\usage{
MomX_conflicts()
}
\description{
This function lists all the conflicts between packages in MomX
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
MomX_conflicts()
}
