% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{MomX_deps}
\alias{MomX_deps}
\title{List all MomX dependencies}
\usage{
MomX_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
MomX packages.}
}
\description{
List all MomX dependencies
}
