% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX-package.R
\name{MomX_funs}
\alias{MomX_funs}
\alias{MomX_attach}
\alias{MomX_detach}
\alias{MomX_update_cran}
\alias{MomX_update_github}
\title{Attach, detach, install, update MomX packages}
\usage{
MomX_attach()

MomX_detach()

MomX_update_cran()

MomX_update_github()
}
\description{
Some MomX packages are still considered experimental.
}
