% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX.R
\name{momx_packages}
\alias{momx_packages}
\title{List all packages in the MomX ecosystem}
\usage{
momx_packages(include_self = FALSE)
}
\arguments{
\item{include_self}{Include MomX package itself? Default is FALSE.}
}
\value{
Character vector of package names
}
\description{
List all packages in the MomX ecosystem
}
\examples{
momx_packages()
}
