% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX.R
\name{momx_status}
\alias{momx_status}
\title{List available MomX packages with status}
\usage{
momx_status()
}
\description{
Show which MomX packages are installed and their versions.
Includes information about packages that don't exist yet.
}
\examples{
momx_status()
}
