% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{MomX_update}
\alias{MomX_update}
\title{Update MomX packages}
\usage{
MomX_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
MomX packages.}
}
\description{
This will check to see if all MomX packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
MomX_update()
}
}
