% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX.R
\name{momx_website}
\alias{momx_website}
\title{Open MomX package websites}
\usage{
momx_website(which = "MomX")
}
\arguments{
\item{which}{Which package website to open? Use "MomX" for the main site,
or any package name like "Momocs2", "Momacs", etc.}
}
\description{
Open the pkgdown website for MomX packages in your browser
}
\examples{
\dontrun{
# Open main MomX website
momx_website()

# Open Momocs2 documentation
momx_website("Momocs2")
}
}
