% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX.R
\docType{package}
\name{MomX-package}
\alias{MomX}
\alias{MomX-package}
\title{MomX: Easily Install and Load the MomX Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The MomX ecosystem is a collection of R packages for morphometrics that work together naturally. This package is designed to make it easy to install and load core MomX packages in a single step. Learn more about MomX at \url{https://momx.github.io/MomX/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://momx.github.io/MomX/}
  \item \url{https://github.com/MomX/MomX}
  \item Report bugs at \url{https://github.com/MomX/MomX/issues}
}

}
\author{
\strong{Maintainer}: Vincent Bonhomme \email{your.email@example.com}

}
\keyword{internal}
