% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX.R
\name{momx_conflicts}
\alias{momx_conflicts}
\title{Conflicts between MomX packages and other packages}
\usage{
momx_conflicts()
}
\value{
A named list of character vectors giving the conflicts
}
\description{
This function lists all conflicts between packages in the MomX ecosystem
and other packages that you have loaded.
}
\examples{
momx_conflicts()
}
