% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomX.R
\name{momx_update}
\alias{momx_update}
\title{Update MomX packages}
\usage{
momx_update(which = c("core", "extended", "all"), from = c("cran", "github"))
}
\arguments{
\item{which}{Which packages to update? "core" (default), "extended", or "all"}

\item{from}{Source to update from: "cran" (default) or "github"}
}
\description{
This will check CRAN and GitHub for updates to MomX packages,
and install them if available. Gracefully handles packages that
don't exist yet.
}
\examples{
\dontrun{
momx_update()
momx_update("all")
momx_update(from = "github")
}
}
